Smooth preset preview

+++ src/Settings.cs
@@ -28,7 +28,7 @@ namespace SnowFlakes
 		public Color SnowdriftsColor = Color.FromArgb(150, Color.LightBlue);
 		public int SnowdriftsResolution = 50;
 		public float SnowdriftsSpeed = 0.2f;
-		public int SnowdriftsDensity = 20;
+		public float SnowdriftsDensity = 20;
 		public int SnowdriftsStart = 40;
 		public int SnowdriftsUpdateDelay = 250;
